fix: add modal component
