+++ src/components/modals/modal/modal.module.css
@@ -1,0 +1,14 @@
+.content{
+    position: relative;
+}
+.close-btn{
+    position: absolute;
+    top: 60px;
+    right: 40px;
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+}
+.close-btn:hover{
+    opacity: 0.6;
+}
